+++ lib/fastlane/plugin/sentry/actions/sentry_create_deploy.rb
@@ -4,7 +4,6 @@ module Fastlane
       def self.run(params)
         require 'shellwords'
 
-        Helper::SentryHelper.check_sentry_cli!
         Helper::SentryConfig.parse_api_params(params)
 
         version = params[:version]
@@ -12,7 +11,6 @@ module Fastlane
         version = "#{version}+#{params[:build]}" if params[:build]
 
         command = [
-          "sentry-cli",
           "releases",
           "deploys",
           version,
@@ -25,7 +23,7 @@ module Fastlane
         command.push('--finished').push(params[:finished]) unless params[:finished].nil?
         command.push('--time').push(params[:time]) unless params[:time].nil?
 
-        Helper::SentryHelper.call_sentry_cli(command)
+        Helper::SentryHelper.call_sentry_cli(params, command)
         UI.success("Successfully created deploy: #{version}")
       end