+++ lib/fastlane/plugin/sentry/actions/sentry_create_release.rb
@@ -4,7 +4,6 @@ module Fastlane
       def self.run(params)
         require 'shellwords'
 
-        Helper::SentryHelper.check_sentry_cli!
         Helper::SentryConfig.parse_api_params(params)
 
         version = params[:version]
@@ -12,14 +11,13 @@ module Fastlane
         version = "#{version}+#{params[:build]}" if params[:build]
 
         command = [
-          "sentry-cli",
           "releases",
           "new",
           version
         ]
         command.push("--finalize") if params[:finalize] == true
 
-        Helper::SentryHelper.call_sentry_cli(command)
+        Helper::SentryHelper.call_sentry_cli(params, command)
         UI.success("Successfully created release: #{version}")
       end