+++ lib/fastlane/plugin/sentry/actions/sentry_upload_dsym.rb
@@ -2,7 +2,6 @@ module Fastlane
   module Actions
     class SentryUploadDsymAction < Action
       def self.run(params)
-        Helper::SentryHelper.check_sentry_cli!
         Helper::SentryConfig.parse_api_params(params)
 
         # Params - dSYM
@@ -16,14 +15,14 @@ module Fastlane
           UI.user_error!("dSYM does not exist at path: #{path}") unless File.exist? path
         end
 
-        command = ["sentry-cli", "upload-dsym"]
+        command = ["upload-dsym"]
         command.push("--symbol-maps") unless params[:symbol_maps].nil?
         command.push(params[:symbol_maps]) unless params[:symbol_maps].nil?
         command.push("--info-plist") unless params[:info_plist].nil?
         command.push(params[:info_plist]) unless params[:info_plist].nil?
         command += dsym_paths
 
-        Helper::SentryHelper.call_sentry_cli(command)
+        Helper::SentryHelper.call_sentry_cli(params, command)
         UI.success("Successfully uploaded dSYMs!")
       end
 
@@ -72,7 +71,7 @@ module Fastlane
                                       optional: true,
                                       verify_block: proc do |value|
                                         UI.user_error! "Could not find Info.plist at path '#{value}'" unless File.exist?(value)
-                                      end)
+                                      end),
         ]
       end