+++ spec/sentry_create_deploy_spec.rb
@@ -2,10 +2,9 @@ describe Fastlane do
   describe Fastlane::FastFile do
     describe "create deploy" do
       it "accepts app_identifier" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         allow(CredentialsManager::AppfileConfig).to receive(:try_fetch_value).with(:app_identifier).and_return(false)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "deploys", "app.idf@1.0", "new", "--env", "staging"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "deploys", "app.idf@1.0", "new", "--env", "staging"]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_create_deploy(
@@ -16,10 +15,9 @@ describe Fastlane do
       end
 
       it "accepts build" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         allow(CredentialsManager::AppfileConfig).to receive(:try_fetch_value).with(:app_identifier).and_return(false)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "deploys", "1.0+123", "new", "--env", "staging"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "deploys", "1.0+123", "new", "--env", "staging"]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_create_deploy(
@@ -30,9 +28,8 @@ describe Fastlane do
       end
 
       it "does not prepend app_identifier if not specified" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "deploys", "1.0", "new", "--env", "staging"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "deploys", "1.0", "new", "--env", "staging"]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_create_deploy(
@@ -42,9 +39,8 @@ describe Fastlane do
       end
 
       it "includes --name if present" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "deploys", "1.0", "new", "--env", "staging", "--name", "fixture-name"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "deploys", "1.0", "new", "--env", "staging", "--name", "fixture-name"]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_create_deploy(
@@ -55,9 +51,8 @@ describe Fastlane do
       end
 
       it "includes --url if present" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "deploys", "1.0", "new", "--env", "staging", "--url", "http://www.sentry.io"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "deploys", "1.0", "new", "--env", "staging", "--url", "http://www.sentry.io"]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_create_deploy(
@@ -68,9 +63,8 @@ describe Fastlane do
       end
 
       it "includes --started and --finished if present" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "deploys", "1.0", "new", "--env", "staging", "--started", 1622630647, "--finished", 1622630700]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "deploys", "1.0", "new", "--env", "staging", "--started", 1622630647, "--finished", 1622630700]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_create_deploy(
@@ -82,9 +76,8 @@ describe Fastlane do
       end
 
       it "includes --started if present" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "deploys", "1.0", "new", "--env", "staging", "--started", 1622630647]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "deploys", "1.0", "new", "--env", "staging", "--started", 1622630647]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_create_deploy(
@@ -95,9 +88,8 @@ describe Fastlane do
       end
 
       it "includes --finished if present" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "deploys", "1.0", "new", "--env", "staging", "--finished", 1622630700]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "deploys", "1.0", "new", "--env", "staging", "--finished", 1622630700]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_create_deploy(
@@ -108,9 +100,8 @@ describe Fastlane do
       end
 
       it "includes --time if present" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "deploys", "1.0", "new", "--env", "staging", "--time", 180]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "deploys", "1.0", "new", "--env", "staging", "--time", 180]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_create_deploy(