+++ spec/sentry_set_commits_spec.rb
@@ -2,10 +2,9 @@ describe Fastlane do
   describe Fastlane::FastFile do
     describe "set commits" do
       it "accepts app_identifier" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         allow(CredentialsManager::AppfileConfig).to receive(:try_fetch_value).with(:app_identifier).and_return(false)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "set-commits", "app.idf@1.0"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "set-commits", "app.idf@1.0"]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_set_commits(
@@ -15,10 +14,9 @@ describe Fastlane do
       end
 
       it "accepts build" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         allow(CredentialsManager::AppfileConfig).to receive(:try_fetch_value).with(:app_identifier).and_return(false)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "set-commits", "1.0+123"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "set-commits", "1.0+123"]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_set_commits(
@@ -28,9 +26,8 @@ describe Fastlane do
       end
 
       it "does not prepend app_identifier if not specified" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "set-commits", "1.0"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "set-commits", "1.0"]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_set_commits(
@@ -39,9 +36,8 @@ describe Fastlane do
       end
 
       it "includes --auto when true" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "set-commits", "1.0", "--auto"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "set-commits", "1.0", "--auto"]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_set_commits(
@@ -51,9 +47,8 @@ describe Fastlane do
       end
 
       it "omits --auto when not present" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "set-commits", "1.0"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "set-commits", "1.0"]).and_return(true)
         Fastlane::FastFile.new.parse("lane :test do
             sentry_set_commits(
               version: '1.0')
@@ -61,9 +56,8 @@ describe Fastlane do
       end
 
       it "omits --auto when false" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "set-commits", "1.0"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "set-commits", "1.0"]).and_return(true)
         Fastlane::FastFile.new.parse("lane :test do
             sentry_set_commits(
               version: '1.0',
@@ -72,9 +66,8 @@ describe Fastlane do
       end
 
       it "includes --clear when true" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "set-commits", "1.0", "--clear"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "set-commits", "1.0", "--clear"]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_set_commits(
@@ -84,9 +77,8 @@ describe Fastlane do
       end
 
       it "omits --clear when not present" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "set-commits", "1.0"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "set-commits", "1.0"]).and_return(true)
         Fastlane::FastFile.new.parse("lane :test do
             sentry_set_commits(
               version: '1.0')
@@ -94,9 +86,8 @@ describe Fastlane do
       end
 
       it "omits --clear when false" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "set-commits", "1.0"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "set-commits", "1.0"]).and_return(true)
         Fastlane::FastFile.new.parse("lane :test do
             sentry_set_commits(
               version: '1.0',
@@ -105,9 +96,8 @@ describe Fastlane do
       end
 
       it "includes --commit when given" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "set-commits", "1.0", "--commit", "abc"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "set-commits", "1.0", "--commit", "abc"]).and_return(true)
 
         Fastlane::FastFile.new.parse("lane :test do
             sentry_set_commits(
@@ -117,9 +107,8 @@ describe Fastlane do
       end
 
       it "omits --commit when not not given" do
-        expect(Fastlane::Helper::SentryHelper).to receive(:check_sentry_cli!).and_return(true)
         expect(Fastlane::Helper::SentryConfig).to receive(:parse_api_params).and_return(true)
-        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(["sentry-cli", "releases", "set-commits", "1.0"]).and_return(true)
+        expect(Fastlane::Helper::SentryHelper).to receive(:call_sentry_cli).with(anything, ["releases", "set-commits", "1.0"]).and_return(true)
         Fastlane::FastFile.new.parse("lane :test do
             sentry_set_commits(
               version: '1.0')