+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- feat: Add sentry cli path as parameter for actions (#97)
+
 ## 1.11.1
 
 ### Fixes